Adding Middlware Factory test
