+++ tests/functional/middleware_factory/required_dependencies/test_middleware_factory.py
@@ -62,37 +62,6 @@ def test_factory_exception_propagation(say_bye_middleware, say_hi_middleware):
         lambda_handler({}, {})
 
 
-def test_factory_explicit_tracing(monkeypatch):
-    monkeypatch.setenv("POWERTOOLS_TRACE_DISABLED", "true")
-
-    @lambda_handler_decorator(trace_execution=True)
-    def no_op(handler, event, context):
-        ret = handler(event, context)
-        return ret
-
-    @no_op
-    def lambda_handler(evt, ctx):
-        return True
-
-    lambda_handler({}, {})
-
-
-def test_factory_explicit_tracing_env_var(monkeypatch):
-    monkeypatch.setenv("POWERTOOLS_TRACE_MIDDLEWARES", "true")
-    monkeypatch.setenv("POWERTOOLS_TRACE_DISABLED", "true")
-
-    @lambda_handler_decorator
-    def no_op(handler, event, context):
-        ret = handler(event, context)
-        return ret
-
-    @no_op
-    def lambda_handler(evt, ctx):
-        return True
-
-    lambda_handler({}, {})
-
-
 def test_factory_decorator_with_kwarg_params(capsys):
     @lambda_handler_decorator
     def log_event(handler, event, context, log_event=False):
